+++ node/clients/users/index.ts
@@ -27,11 +27,14 @@ class UserDataSource extends IODataSource {
     )
   }
 
-  public getTotalNumber = async (vtexIdToken: string) => {
-    const res = await this.http.getRaw(`${DATA_ENTITY}/search`, {
-      metric: 'crm-get-users',
-      headers: withAuthToken({ 'REST-Range': 'resources=1-1' })(vtexIdToken),
-    })
+  public getTotalNumber = async (filter: string, vtexIdToken: string) => {
+    const res = await this.http.getRaw(
+      `${DATA_ENTITY}/search?_fields=_all${filter ? `&${filter}` : ''}`,
+      {
+        metric: 'crm-get-users',
+        headers: withAuthToken({ 'REST-Range': 'resources=1-1' })(vtexIdToken),
+      }
+    )
 
     const contentRange = res.headers['rest-content-range']
     return parseInt(contentRange.split('/')[1])